+++ src/main/java/edu/ie3/datamodel/io/sink/CsvFileSink.java
@@ -10,6 +10,7 @@ import edu.ie3.datamodel.io.connectors.CsvFileConnector;
 import edu.ie3.datamodel.io.csv.BufferedCsvWriter;
 import edu.ie3.datamodel.io.csv.CsvFileDefinition;
 import edu.ie3.datamodel.io.extractor.Extractor;
+import edu.ie3.datamodel.io.extractor.HasEm;
 import edu.ie3.datamodel.io.extractor.NestedEntity;
 import edu.ie3.datamodel.io.naming.FileNamingStrategy;
 import edu.ie3.datamodel.io.processor.ProcessorProvider;
@@ -225,14 +226,29 @@ public class CsvFileSink implements InputDataSink, OutputDataSink {
             .flatMap(Optional::stream)
             .collect(Collectors.toSet());
 
+    // extract all EmInput entities
+    Set<EmInput> emInputs =
+        systemParticipants.allEntitiesAsList().stream()
+            .filter(Objects::nonNull)
+            .map(HasEm.class::cast)
+            .map(HasEm::getControllingEm)
+            .flatMap(Optional::stream)
+            .flatMap(
+                em ->
+                    Stream.iterate(
+                        em, Objects::nonNull, current -> current.getControllingEm().orElse(null)))
+            .collect(Collectors.toSet());
+
     // persist all entities
     Stream.of(
             rawGridElements.allEntitiesAsList(),
             systemParticipants.allEntitiesAsList(),
             graphicElements.allEntitiesAsList(),
             types,
-            operators)
+            operators,
+            emInputs)
         .flatMap(Collection::stream)
+        .collect(Collectors.toSet())
         .forEach(this::persistIgnoreNested);
   }